jobs/epic-cron/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
pytest-mock
requests
pyhamcrest
pytest-cov

# Lint and code style
flake8==4.0.1
flake8-blind-except
flake8-debugger
flake8-docstrings
flake8-isort
flake8-quotes
pep8-naming
autopep8
coverage
pylint==v2.13.8
pylint-flask
FreezeGun

# docker
lovely-pytest-docker
pytest-asyncio